+++ rrr-backend/src/main/java/com/nous/rollingrevenue/service/impl/ProbabilityTypeServiceImpl.java
@@ -3,6 +3,7 @@ package com.nous.rollingrevenue.service.impl;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -53,12 +54,14 @@ public class ProbabilityTypeServiceImpl implements ProbabilityTypeService {
 	public void deleteProbabilityTypeById(Long probabilityTypeId) {
 		List<RevenueEntry> revenueEntryList = revenueEntryRespository.findByProbabilityTypeId(probabilityTypeId);
 		if (!revenueEntryList.isEmpty()) {
-			throw new RecordNotFoundException(
-					"ProbabilityType is already linked to RevenueEntry");
+			throw new RecordNotFoundException("ProbabilityType is already linked to RevenueEntry");
 		}
-		probabilityTypeRepository.findById(probabilityTypeId)
-				.orElseThrow(() -> new RecordNotFoundException(ErrorConstants.RECORD_NOT_EXIST + probabilityTypeId));
-		probabilityTypeRepository.deleteById(probabilityTypeId);
+		Optional<ProbabilityType> findById = probabilityTypeRepository.findById(probabilityTypeId);
+		if (findById.isEmpty()) {
+			throw new RecordNotFoundException(ErrorConstants.RECORD_NOT_EXIST + probabilityTypeId);
+		} else {
+			probabilityTypeRepository.deleteById(probabilityTypeId);
+		}
 	}
 
 	@Override
